Move some methods with parameters from MightyGrid::Base to MightyGrid::Parameters
