+++ lib/mighty_grid.rb
@@ -6,6 +6,7 @@ module MightyGrid
   extend ActiveSupport::Autoload
 
   autoload :Base
+  autoload :Parameters
   autoload :Filters
   autoload :Column
   autoload :GridRenderer
